quick-fix for building Routes without the need of Range

+++ src/main/kotlin/com/colivery/engine/SearchController.kt
@@ -37,7 +37,7 @@ class SearchController {
     @PostMapping("/route/orders")
     fun routeOrders(@RequestBody @Valid request: RouteRequest): RouteResponse {
         val orders = orderService.fetchOrdersByIds(request.orderIds)
-        val range = distanceService.calculateRange(orders, request.coordinate, (request.range ?: 5.0) as Double)
+        val range = distanceService.calculateRange(orders, request.coordinate, (request.range ?: 5.0))
         val allPoIs = poiService.findAllPoIs(request.coordinate, range, orders)
         val activitySequence = routeService.buildRoute(request.coordinate, orders, allPoIs)
